refactor: add useRedirect hook
