+++ src/pages/Register/Register.tsx
@@ -1,31 +1,15 @@
-import { ReactElement, useEffect } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { ReactElement } from 'react';
+import { Link } from 'react-router-dom';
 import { RegisterForm } from '../../components/RegisterForm';
 import { ErrorMessage } from '../../components/shared/ErrorMessage';
 import { PATH } from '../../router/constants/paths';
-import { selectAuthError, selectIsAuthorized } from '../../store/auth/selectors';
 import { Page } from '../../router/types';
-import { useAppDispatch } from '../../store/store';
-import { resetError } from '../../store/auth/slice';
+import { useRedirect } from '../../hooks';
 
 import styles from './register.module.scss';
 
 export default function Register(): ReactElement {
-  const navigate = useNavigate();
-  const isAuthorized = useSelector(selectIsAuthorized);
-  const error = useSelector(selectAuthError);
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    if (isAuthorized) {
-      navigate(PATH[Page.Home]);
-    }
-
-    return (): void => {
-      dispatch(resetError());
-    };
-  }, [isAuthorized, navigate, dispatch]);
+  const error = useRedirect();
 
   return (
     <div className={styles.register}>
